Changed layout / design
Changed the design, etc.

Make SURE to pull down first.  For some reason it's saying I have to
check in all of the resource files.

Pull before you start working on removing the first digits.

Also, when removing the first two digits of the main time, you need to
adjust the width of the text box to get it to show up properly (i.e.
sized correctly)